requirements.txt
asgiref==3.7.2
certifi==2023.7.22
Django==4.2.3
django-filter==23.2
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-ses==3.5.0
django-storages==1.14.2
djangorestframework==3.14.0
jmespath==1.0.1
kafka-python==2.0.2
phonenumbers==8.13.17
Pillow==10.0.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.6
